+++ app/controllers/issues_controller.rb
@@ -42,7 +42,9 @@ class IssuesController < ApplicationController
 
     @issue = @project.backlog_items.build(params[:issue])
     @issue.reported_by = current_user
-    @issue.save
+    if @issue.save
+      IssueMailer.send_issue_raised(@issue.id)
+    end
 
     respond_with(@issue, :location => redirect_to_path || show_path)
   end