+++ app/models/issue.rb
@@ -61,6 +61,7 @@ class Issue < ActiveRecord::Base
     else
       remove_from_backlog
     end
+    IssueMailer.delay.issue_closed(self.id)
   end
 
   # Callback fired afer ticket reopened
@@ -70,6 +71,7 @@ class Issue < ActiveRecord::Base
     else
       add_to_backlog
     end
+    IssueMailer.delay.issue_reopened(self.id)
   end
 
   private