+++ config/initializers/omniauth.rb
@@ -1,3 +1,7 @@
 Rails.application.config.middleware.use(OmniAuth::Builder) do
-  provider :github, ENV['GITHUB_KEY'], ENV['GITHUB_SECRET']
+  begin
+    provider :github, APP_CONFIG.omniauth.github.client_id, APP_CONFIG.omniauth.github.client_secret
+  rescue
+    raise "Please add your github API credentials to config/tyne.yml"
+  end
 end