+++ db/migrate/20130319200745_add_notification_email_to_users.rb
@@ -1,0 +1,5 @@
+class AddNotificationEmailToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :notification_email, :string
+  end
+end