+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.6.2
+
+* Growable Textareas enabled (projects, issues)
+
+## v0.6.1
+
+* Asynchronous delivery of emails
+
+## v0.6.0
+
+* Sends email notifications for certain events (issue raised, closed, reopened)
+
 ## v0.5.1
 
 * Redirect to requested url after successful authentication